requirements.txt
moonshine-voice>=0.0.51
sounddevice
pystray
Pillow
keyboard
pynput
pyperclip
